backend/auditlogs/views.py
from rest_framework import viewsets, permissions, filters
from django_filters.rest_framework import DjangoFilterBackend
from django_filters import rest_framework as django_filters
from .models import AuditLog
from .serializers import AuditLogSerializer

class IsSuperAdmin(permissions.BasePermission):
    """
    Allows access only to superusers.
    """
    def has_permission(self, request, view):
        if not request.user or not request.user.is_authenticated:
            return False
        return bool(request.user.is_superuser or request.user.role == 'SUPERADMIN')

from rest_framework.pagination import PageNumberPagination

class AuditLogPagination(PageNumberPagination):
    page_size = 20
    page_size_query_param = 'page_size'
    max_page_size = 100

class AuditLogFilter(django_filters.FilterSet):
    start_date = django_filters.DateFilter(field_name="timestamp__date", lookup_expr='gte')
    end_date = django_filters.DateFilter(field_name="timestamp__date", lookup_expr='lte')

    class Meta:
        model = AuditLog
        fields = ['action', 'entity_type', 'status', 'actor_userid', 'actor_employee_id', 'start_date', 'end_date']

class AuditLogViewSet(viewsets.ReadOnlyModelViewSet):
    queryset = AuditLog.objects.all().order_by('-timestamp')
    serializer_class = AuditLogSerializer
    permission_classes = [permissions.IsAuthenticated, IsSuperAdmin]
    pagination_class = AuditLogPagination
    filter_backends = [DjangoFilterBackend, filters.SearchFilter]
    
    filterset_class = AuditLogFilter
    # Enhanced search to include actor's name/email and user details
    search_fields = [
        'actor_userid', 
        'actor_employee_id', 
        'entity_type', 
        'details',
        'actor__first_name',
        'actor__last_name',
        'actor__full_name',
        'actor__email'
    ]

    def get_queryset(self):
        print("DEBUG: AuditLogViewSet.get_queryset called")
        qs = super().get_queryset()
        print(f"DEBUG: Found {qs.count()} logs")
        return qs
