backend/requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.4
brotli==1.2.0
cbor2==5.8.0
celery==5.4.0
certifi==2026.1.4
cffi==2.0.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
cssselect2==0.8.0
daphne==4.2.1
Django==4.2.11
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.26.4
drf-yasg==1.21.10
et_xmlfile==2.0.0
fonttools==4.53.0
freetype-py==2.5.1
gunicorn==23.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
lxml==6.0.2
msgpack==1.1.2
nepali-datetime==1.0.8.4
numpy==1.26.4
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.0
pandas==2.2.3
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
q
pycparser==2.23
pydyf==0.11.0
pyHanko==0.33.0
pyhanko-certvalidator==0.29.1
PyJWT==2.11.0
pyOpenSSL==25.3.0
pypdf==6.7.0
pyphen==0.17.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
redis==5.2.1
referencing==0.36.2
reportlab==4.4.9
requests==2.32.5
rpds-py==0.27.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.4.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
webencodings==0.5.1
zope.interface==8.0.1
zopfli==0.2.3.post1
pyaxmlparser==0.3.30
